update product service

+++ src/main/java/com/example/nosqllab1/models/Product.java
@@ -4,7 +4,10 @@ package com.example.nosqllab1.models;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 public class Product {
+
     @Setter
     @Getter
     private String id;
@@ -15,7 +18,7 @@ public class Product {
 
     @Setter
     @Getter
-    private double price;
+    private BigDecimal price;
 
     @Setter
     @Getter
